Move test util.js and template tests out.
