+++ generators/polymer-app/templates/test/util.js
@@ -1,5 +1,5 @@
 ﻿module.exports = {
 	request: require('supertest'),
 	expect: require('chai').expect,
-	server: require('../tmp/build/server/server.js')()
+	server: require('./tmp/build/server/server.js')()
 };
